Service for database log added

+++ backend/r2d2/src/main/java/com/unitybars/r2d2/dao/TaskCheckLogDao.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface TaskCheckLogDao {
     long insertTaskCheckLog(TaskCheckLog taskCheckLog);
 
-    TaskCheckLog getTaskCheckLogById(int id);
+    TaskCheckLog getTaskCheckLogById(long id);
 
-    List<TaskCheckLog> getTaskCheckLogsForServiceCheckLog(int serviceCheckLogId);
+    List<TaskCheckLog> getTaskCheckLogsForServiceCheckLog(long serviceCheckLogId);
 }
